requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asyncssh==2.7.0
    # via scrapli-asyncssh
cffi==1.14.6
    # via cryptography
cryptography==3.4.7
    # via asyncssh
pycparser==2.20
    # via cffi
pyyaml==5.4.1
    # via -r requirements.in
scrapli-asyncssh==2020.12.23
    # via -r requirements.in
scrapli==2021.1.30
    # via
    #   -r requirements.in
    #   scrapli-asyncssh
